backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asn1crypto==1.5.1
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.6.0
cryptography==41.0.7
dill==0.3.7
flake8==7.0.0
Flask==3.0.0
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.17.3
google-cloud-core==2.4.1
google-cloud-datastore==2.15.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pg8000==1.30.4
platformdirs==4.1.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.25
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==1.26.18
uuid==1.30
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
